Commit Develop
Separação do arquivo principal em mais 3 arquivos .c e 3 arquivos .h contendo cada uma das 3 funções criadas.
