+++ Somar.h
@@ -1,0 +1,8 @@
+#ifndef SOMAR_H
+#define SOMAR_H
+
+// Definição dos elementos da função 1
+
+double somar(double array[], int tamanho);
+
+#endif
