+++ maximoValor.h
@@ -1,0 +1,8 @@
+#ifndef MAXIMOVALOR_H
+#define MAXIMOVALOR_H
+
+//definição dos elementos da função 2
+
+double maximoValor(double array[], int tamanho);
+
+#endif
