+++ media.h
@@ -1,0 +1,8 @@
+#ifndef MEDIA_H
+#define MEDIA_H
+
+// Definição dos valores da função 3
+
+double media(double array[], int tamanho);
+
+#endif
